content manager service, sync doc and link delete

+++ api/src/services/LinkService.ts
@@ -1,9 +1,28 @@
 import { getCustomRepository, getTreeRepository } from 'typeorm'
-import { LinkRepository } from '../../repositories/LinkRepository'
-import { Link } from '../../entities/Link'
-import { LinkCreateValue, LinkUpdateValue } from '../../values/link'
+import { LinkRepository } from '../repositories/LinkRepository'
+import { Link } from '../entities/Link'
+import { LinkCreateValue, LinkUpdateValue } from '../values/link'
+import { clientError } from '../errors/httpError'
+import { ContentManager } from './content/ContentManager'
 
 export class LinkService {
+  private contentManager: ContentManager
+
+  
+  private static instance: LinkService
+
+  private constructor() {
+    this.contentManager = ContentManager.getInstance()
+  }
+
+  static getInstance() {
+    if (!LinkService.instance) {
+      LinkService.instance = new LinkService()
+    }
+
+    return LinkService.instance
+  }
+
   getLinkRepository(): LinkRepository {
     return getCustomRepository(LinkRepository)
   }
@@ -45,9 +64,19 @@ export class LinkService {
     return this.getLinkRepository().update(id, data.toObject())
   }
 
-  delete(id: number) {
-    return this.getLinkRepository().delete({
+  async delete(id: number) {
+    const link = await this.getLinkById(id)
+
+    if (!link) {
+      throw clientError('Error deleting link')
+    }
+
+    const res = await this.getLinkRepository().delete({
       id,
     })
+
+    this.contentManager.linkDeleted(link)
+
+    return res
   }
 }
